+++ frontend/src/_configs/base-configs.jsx
