+++ frontend/src/_configs/index.jsx
@@ -1,1 +1,1 @@
-export * from './base_configs'
+export * from './base-configs'
