+++ frontend/src/_constants/index.jsx
@@ -1,2 +1,2 @@
-export * from './user_constants'
-export * from './alert_constants'
+export * from './user-constants'
+export * from './alert-constants'
