+++ frontend/src/_reducers/index.jsx
@@ -1,6 +1,6 @@
 import { combineReducers } from 'redux'
-import authentication from './authentication_reducer'
-import alert from './alert_reducer'
+import authentication from './authentication-reducer'
+import alert from './alert-reducer'
 
 const rootReducer = combineReducers({
   authentication,
