+++ frontend/src/_services/index.jsx
@@ -1,1 +1,1 @@
-export * from './user_services'
+export * from './user-services'
